Refac: Move Message Classes out of group-Screen file
